Services buy (#15)

+++ api/src/main/java/com/findyourguide/api/service/ServiceServiceImpl.java
@@ -5,15 +5,16 @@ import com.findyourguide.api.dto.service.ServiceDTO;
 import com.findyourguide.api.dto.service.UpdateServiceDTO;
 import com.findyourguide.api.entity.Guide;
 import com.findyourguide.api.entity.Service;
+import com.findyourguide.api.error.ServiceNotFoundException;
+import com.findyourguide.api.error.UserNotFoundException;
 import com.findyourguide.api.mapper.ServiceMapper;
 import com.findyourguide.api.repository.GuideRepository;
 import com.findyourguide.api.repository.ServiceRepository;
+import com.findyourguide.api.service.interfaces.IServiceService;
 import com.findyourguide.api.util.Populate;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -25,43 +26,37 @@ public class ServiceServiceImpl implements IServiceService {
     private final GuideRepository guideRepository;
 
     public ServiceDTO create(CreateServiceDTO serviceDTO) {
-        Service service = new Service();
-        service.setName(serviceDTO.getName());
-        service.setDate(serviceDTO.getDate());
-        service.setServiceType(serviceDTO.getServiceType());
-        service.setPrice(serviceDTO.getPrice());
-        service.setRate(0);
-        service.setQuantity(serviceDTO.getQuantity());
-        service.setDescription(serviceDTO.getDescription());
-        service.setCountry(serviceDTO.getCountry());
-        service.setCity(serviceDTO.getCity());
-
-        Optional<Guide> guide = guideRepository
-                .findUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
-
-        guide.ifPresent(service::setGuide);
+        Guide guide = guideRepository
+                .findUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(() -> new UserNotFoundException());
+        Service service = ServiceMapper.mapToServiceEntityFromCreateServiceDTO(serviceDTO, guide);
         serviceRepository.save(service);
-        return ServiceMapper.toDTO(service);
+        return ServiceMapper.mapToServiceDTO(service);
     }
 
     public List<ServiceDTO> findAll() {
         return serviceRepository.findAll().stream().map(
-                ServiceMapper::toDTO).collect(Collectors.toList());
+                ServiceMapper::mapToServiceDTO).collect(Collectors.toList());
     }
 
     public List<ServiceDTO> findAllByGuide(Long id) {
         return guideRepository.findById(id)
                 .map(guide -> guide.getGuideServices().stream()
-                        .map(ServiceMapper::toDTO)
+                        .map(ServiceMapper::mapToServiceDTO)
                         .collect(Collectors.toList()))
-                .orElse(Collections.emptyList());
+                .orElseThrow(() -> new UserNotFoundException());
     }
 
-    public Optional<Service> findById(Long id) {
-        return serviceRepository.findById(id);
+    public ServiceDTO findById(Long id) throws ServiceNotFoundException {
+        return serviceRepository.findById(id)
+                .map(ServiceMapper::mapToServiceDTO)
+                .orElseThrow(() -> new ServiceNotFoundException());
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws ServiceNotFoundException {
+        if (!serviceRepository.existsById(id)) {
+            throw new ServiceNotFoundException();
+        }
         serviceRepository.deleteById(id);
     }
 
@@ -71,8 +66,10 @@ public class ServiceServiceImpl implements IServiceService {
             Service updatedService = Populate.populateUpdateService(serviceOptional.get(), updateServiceDTO);
             System.out.println(updatedService);
             serviceRepository.save(updatedService);
-            return ServiceMapper.toDTO(updatedService);
+            return ServiceMapper.mapToServiceDTO(updatedService);
+        } else {
+            throw new ServiceNotFoundException();
         }
-        return null;
+
     }
 }
